Move optimizers/transformer_primitives.py to transformers/transformer_primitives.py
