+++ cirq-core/cirq/__init__.py
@@ -337,11 +337,6 @@ from cirq.optimizers import (
     EjectZ,
     ExpandComposite,
     is_negligible_turn,
-    map_moments,
-    map_operations,
-    map_operations_and_unroll,
-    merge_moments,
-    merge_operations,
     merge_single_qubit_gates_into_phased_x_z,
     merge_single_qubit_gates_into_phxz,
     MergeInteractions,
@@ -360,6 +355,14 @@ from cirq.optimizers import (
     two_qubit_matrix_to_diagonal_and_operations,
     two_qubit_matrix_to_sqrt_iswap_operations,
     three_qubit_matrix_to_operations,
+)
+
+from cirq.transformers import (
+    map_moments,
+    map_operations,
+    map_operations_and_unroll,
+    merge_moments,
+    merge_operations,
     unroll_circuit_op,
     unroll_circuit_op_greedy_earliest,
     unroll_circuit_op_greedy_frontier,
