+++ cirq-core/cirq/transformers/transformer_primitives.py
